requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
bottlenose==1.1.2
certifi==2017.4.17
chardet==3.0.4
decorator==4.0.11
enum34==1.1.6
fire==0.1.1
idna==2.5
ipython==5.4.1
ipython-genutils==0.2.0
lxml==3.8.0
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
python-amazon-simple-product-api==2.2.11
python-dateutil==2.6.0
requests==2.17.3
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
slackclient==1.0.5
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
websocket-client==0.40.0
